fixed admin

+++ resources/views/layouts/feed.blade.php
@@ -11,16 +11,18 @@
                                     <div class="col-xs-4 text-left"><b>{{ $set->name }}</b></div>
                                     <div class="col-xs-4 text-left">by {{ $set->user->name }}</div>
                                     <div class="col-xs-4 text-left">
-                                    @if(in_array($set->id, $followed_sets))
-                                        {{ Form::open(['method' => 'delete', 'route' => ['studies.destroy', $set->id]]) }}
-                                            {!! Form::hidden('sid', $set->id), null !!}
-                                            {!! Form::submit('Unstudy', ['class' => 'btn btn-warning btn-block']) !!}
-                                        {{ Form::close() }}
-                                    @else
-                                        {{ Form::open(['method' => 'post', 'route' => 'studies.store']) }}
-                                            {!! Form::hidden('sid', $set->id), null !!}
-                                            {!! Form::submit('Study', ['class' => 'btn btn-primary btn-block']) !!}
-                                        {{ Form::close() }}
+                                    @if(!auth()->user()->isAdmin())
+                                        @if(in_array($set->id, $followed_sets))
+                                            {{ Form::open(['method' => 'delete', 'route' => ['studies.destroy', $set->id]]) }}
+                                                {!! Form::hidden('sid', $set->id), null !!}
+                                                {!! Form::submit('Unstudy', ['class' => 'btn btn-warning btn-block']) !!}
+                                            {{ Form::close() }}
+                                        @else
+                                            {{ Form::open(['method' => 'post', 'route' => 'studies.store']) }}
+                                                {!! Form::hidden('sid', $set->id), null !!}
+                                                {!! Form::submit('Study', ['class' => 'btn btn-primary btn-block']) !!}
+                                            {{ Form::close() }}
+                                        @endif
                                     @endif
                                     </div>
                                 </p>
